+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 🚀 Features
 
 - Allow special characters in quoted names in `use line` expressions from Fortran namelist-based lattice files created by Elegant and Bmad (see #637) (@cr-xu)
+- Add an optional, serializable `metadata` dictionary to all elements (and segments) for storing annotations such as control-system PV names that are not used in simulation. The field is preserved by `clone` and through LatticeJSON save/load. (see #645) (@cr-xu)
 
 ### 🐛 Bug fixes